requirements.txt
numpy==1.18.1
matplotlib==3.1.3
pandas==1.0.1
streamlit==0.61.0
Pillow==7.1.2
scikit_learn==0.23.1
